Moved getStorageLevelOfRDD() to a TestUtils class, since it's no longer used in non-test code.
